Paging!
I couldn't seem to figure out Google's Paging library DataSources so I made my own basic version of a paging implementation.
As the RecyclerView scrolls, it tracks how far it has scrolled. Once it passes 75%, it tries to load the next set of data.
When new data comes in, it is added to the data in the Adapter and the data set is updated.

+++ app/src/main/java/com/mathewsmobile/spacebook/model/LoginResponse.kt
@@ -6,7 +6,7 @@ data class LoginResponse(val status: String, val data: LoginData?, val error: Lo
         return data != null && error == null
     }
     
-    data class LoginData(val id: Long, val email: String, val name: String, val registeredAt: String, val githubUsername: String?, val rating: Double)
+    data class LoginData(val id: Int, val email: String, val name: String, val registeredAt: String, val githubUsername: String?, val rating: Double)
     
     data class LoginError(val type: String)
 }
